Redo kerneltest-harness
This is a sweeping rewrite of kerneltest-harness. It changes the
database schema with no attempt at a migration path, moves from OpenID
to OpenID Connect, offers an HTTP API authenticated with OpenID Connect,
significantly refactors the entire code base, and adds CI via Tox and
Travis. This definitely breaks the "one change per commit" rule, but I
don't think it's worth the effort to do it that way.

Signed-off-by: Jeremy Cline <jcline@redhat.com>